API

The following methods are exposed when using JDR.

Processing RPKI files

JDR.RPKI.process_tas — Function

process_tas([tal_urls::Dict]; kw...)

Process all trust anchors configured in JDR.toml. This is likely the most common way to start doing anything using JDR.jl:

using JDR;
(tree, lookup) = process_tas()

Optionally, a Dict of similar form can be passed directly to specify trust anchors not in the configuration file, or a subset of those that are specified.

Optional keyword arguments:

repodir::String – defaults to CFG["rpki"]["rsyncrepo"], i.e. the JDR.toml config file.

stripTree::Bool – drop the ASN.1 tree from objects after validation, defaults to false

nicenames::Bool – enrich the ASN.1 tree with human-readable fieldnames, defaults to true

Returns Tuple{RPKINode,Lookup}

JDR.RPKI.process_ta — Function

process_ta(ta_cer_fn::String; kw...)

Takes a first cerficate to parse and processes all related descending files. Called by process_tas.

Optional keyword arguments:

repodir::String – defaults to CFG["rpki"]["rsyncrepo"], i.e. the JDR.toml config file. Useful for processing data that is stored in a non-default directory structure, for example when the TA certificate is stored in a different directory than the RPKI files.

lookup – defaults to an empty Lookup()

stripTree::Bool – drop the ASN.1 tree from objects after validation, defaults to false

nicenames::Bool – enrich the ASN.1 tree with human-readable fieldnames, defaults to true

Returns Tuple{RPKINode,Lookup}

Search based on RPKINode tree and Lookup

JDR.RPKICommon.search — Function

search(l::Lookup, asn::AutSysNum)

Search for RPKINode's related to a AutSysNum

search(l::Lookup, filename::AbstractString)

Search for RPKINode's related to filename

search(tree::RPKINode, ipr::IPRange{T}, include_more_specific::Bool=false)

Find RPKINodes holding CERs or ROAs containing resources queried for. ROAs are matched if the ipr matches resources on the EE, or in the VRPs.

search(tree::RPKINode, id::IssuerSubject)

Search certificates containing this issuer/subject ID

JDR.RPKICommon.new_since — Function

new_since(tree::RPKINode, tp::TimePeriod=Hour(1)) :: Vector{RPKINode}

Find files published in the last tp, defaulting to one hour.

For CER objects, this is based on .notBefore. For MFT and CRL, this is based on .this_update.

Helpers

JDR.RPKICommon.get_object — Function

get_object(n::RPKINode)
get_object(o::RPKIObject)

Returns the wrapped object (CER, MFT, CRL, ROA) for an RPKINode or RPKIObject.

JDR.RPKICommon.get_pubpoint — Function

get_pubpoint(node::RPKINode) :: String

Get the domain name of the publication point for the file represented by this RPKINode.

JDR.RPKICommon.print_ASN1 — Function

print_ASN1(n::RPKINode{T}; max_lines=0)
print_ASN1(o::RPKIObject{T}; max_lines=0)

Print the annotated ASN.1 tree structure for

JDR.Common.prefixlen — Function

prefixlen(r::IPRange{IPv6})
prefixlen(r::IPRange{IPv4})

Returns the prefix length of an IPRange.

JDR.RPKICommon.vrps — Function

vrps(r::ROA)
vrps_v6(r::ROA)
vrps_v4(r::ROA)

Get the VRPs listed on a ROA in a Vector{VRP}

Index

JDR.Common.prefixlen

JDR.RPKI.process_ta

JDR.RPKI.process_tas

JDR.RPKICommon.get_object

JDR.RPKICommon.get_pubpoint

JDR.RPKICommon.new_since

JDR.RPKICommon.print_ASN1

JDR.RPKICommon.search

JDR.RPKICommon.vrps