Type reference

Tree structure

Types constituting the tree as returned by process_tas and process_ta.

JDR.RPKICommon.RPKINode — Type

RPKINode

Represents a file in the RPKI, with pointers to its parent, children and possible siblings.

The obj points to the RPKIObject{T} of this RPKINode, e.g. a RPKIObject{CER}.

Fields:

parent::Union{Nothing, RPKINode}

children::Vector{RPKINode}

siblings::Union{Nothing, Vector{RPKINode}}

obj::Union{Nothing, RPKIObject}

remark_counts_me::Union{Nothing, RemarkCounts_t}

remark_counts_children::Union{Nothing, RemarkCounts_t}

JDR.RPKICommon.RPKIObject — Type

RPKIObject{T<:RPKIFile}

Contains an object T (i.e. a CER, MFT, CRL or ROA), decoded from filename into an annotated ASN1 tree in tree. Any warnings or errors for this object are stored in remarks.

Fields:

filename::String

tree::Union{Nothing, Node} – ASN1.Node

object::T

remarks::Union{Nothing, Vector{Remark}}

remarks_tree::Union{Nothing, Vector{Remark}}

sig_valid::Union{Nothing, Bool}

cms_digest_valid::Union{Nothing, Bool}

JDR.RPKICommon.RPKIFile — Type

Abstract type to parameterize RPKIObject

julia> subtypes(RPKIFile)
5-element Vector{Any}:
 CER
 CRL
 MFT
 ROA
 RootCER

Specific RPKIObject types

JDR.RPKICommon.RootCER — Type

RootCER <: RPKIFile

Synthetic, 'empty certificate' to act as the single entry point to the RPKINode graph, with all the RIR trust anchor certificates being its children.

JDR.RPKICommon.CER — Type

CER <: RPKIFile

Represents a decoded certificate (.cer) file.

Fields:

Extracted from decoded file:

serial – Serial number on this certificate

notBefore/notAfter – DateTime fields

pubpoint – String

manifest – String

rrdp_notify – String

selfsigned – Bool or Nothing

issuer – String

subject – String

Set after validation:

validsig – Bool or Nothing

resources_valid – Bool or Nothing

JDR.RPKICommon.MFT — Type

MFT <: RPKIFile

Represents a decoded manifest (.mft) file.

Fields:

Extracted from decoded file:

files – Vector{String} containing filenames listed on this manifest

loops – Vector{String} or Nothing listing the filenames causing a loop

missing_files – Vector{String} of filenames listed in the manifest but not found on disk

this_update: DateTime

next_update: DateTime

JDR.RPKICommon.CRL — Type

CRL <: RPKIFile

Represents a decoded Certificate Revocation List (.crl) file.

Fields:

revoked_serials – Vector of SerialNumber's

this_update – DateTime

next_update – DateTime

JDR.RPKICommon.ROA — Type

ROA <: RPKIFile

Represents a decoded Route Origin Authorization (.roa) file.

Fields:

asid – Integer

vrp_tree – VRPS

resources_v6 – IntervalTree{IPv6} of the IPv6 resources in the EE certificate

resources_v4 – IntervalTree{IPv4} of the IPv4 resources in the EE certificate

After validation:

resources_valid – Bool or Nothing

The Lookup struct

JDR.RPKICommon.Lookup — Type

Lookup

Provides mappings/shortcuts to RPKINodes of particular interest.

Fields:

ASNs::Dict{AutSysNum}{Vector{RPKINode}}

filenames::Dict{String}{RPKINode}

missing_files::Dict{String}{RPKINode}

resources_v6::IntervalTree{IPv6,IntervalValue{IPv6,RPKINode}}

resources_v4::IntervalTree{IPv4,IntervalValue{IPv4,RPKINode}}

pubpoints::Dict{String}{Pair{Int,Set{RPKINode}}}

too_specific::Vector{RPKINode}

invalid_signatures::Vector{RPKIObject{T} where T}

invalid_certs::Vector{RPKINode}

valid_certs::Vector{RPKINode}

Common types

JDR.Common.AutSysNum — Type

Wrapper for asn::UInt32. Can be constructed directly via any Integer, or be parsed from a String:

AutSysNum(s::AbstractString)

JDR.Common.IPRange — Type

IPRange{T<:IPAddr}

Describes an IP prefix or range by only storing the first and last Sockets.IPAddr.

JDR.RPKICommon.VRPS — Type

Holds the IPv6 and IPv4 resources listed on this ROA, as an IntervalTree, with Values denoting the maxlength.

Fields:

resources_v6::_VRPS{IPv6}

resources_v4::_VRPS{IPv4}

These field are named this way so we can easily to coverage checks between different RPKIObject types. _VRPS is an alias for IntervalMap{T, UInt8}.

JDR.RPKICommon.SerialNumber — Type

Simple wrapper for serial::Integer

Index

JDR.Common.AutSysNum

JDR.Common.IPRange

JDR.RPKICommon.CER

JDR.RPKICommon.CRL

JDR.RPKICommon.Lookup

JDR.RPKICommon.MFT

JDR.RPKICommon.ROA

JDR.RPKICommon.RPKIFile

JDR.RPKICommon.RPKINode

JDR.RPKICommon.RPKIObject

JDR.RPKICommon.RootCER

JDR.RPKICommon.SerialNumber

JDR.RPKICommon.VRPS